Redirect auf Login-Seite, wenn keine gültige Session besteht

+++ ci/app/Controllers/Persons.php
@@ -23,6 +23,8 @@ class Persons extends BaseController
 
     public function index()
     {
+        if (empty($this->session->get("sessionUserId"))) return redirect()->to(base_url("./login"));
+
         $headData['title'] = 'Mitglieder';
         $headData['heading'] = 'Aufgabenplaner: Mitglieder';
         $data['personen'] = $this->mitgliedModel->getMitglieder();
@@ -49,6 +51,8 @@ class Persons extends BaseController
 
     public function save()
     {
+        if (empty($this->session->get("sessionUserId"))) return redirect()->to(base_url("./login"));
+
         if (empty($_POST['userId']))
         {
             // Prüfe, dass alle Felder ausgefüllt sind
@@ -90,6 +94,8 @@ class Persons extends BaseController
 
     public function delete()
     {
+        if (empty($this->session->get("sessionUserId"))) return redirect()->to(base_url("./login"));
+
         if (!empty($_GET['id'])) {
             $this->mitgliedModel->deleteMitglied($_GET['id']);
         }
